+++ Package.swift
@@ -9,6 +9,7 @@ let package = Package(
     products: [
         .library(name: "Danger", type: .dynamic, targets: ["Danger"]),
         .library(name: "DangerFixtures", type: .dynamic, targets: ["DangerFixtures"]),
+        .library(name: "DangerDeps", type: .dynamic, targets: ["Danger-Swift"]), // dev
         .executable(name: "danger-swift", targets: ["Runner"]),
     ],
     dependencies: [
@@ -21,10 +22,12 @@ let package = Package(
         .package(url: "https://github.com/orta/Komondor", from: "1.0.0"), // dev
         .package(url: "https://github.com/nicklockwood/SwiftFormat", from: "0.35.8"), // dev
         .package(url: "https://github.com/Realm/SwiftLint", from: "0.28.1"), // dev
-        .package(url: "https://github.com/f-meloni/Rocket", from: "0.3.2"), // dev
         .package(url: "https://github.com/pointfreeco/swift-snapshot-testing.git", from: "1.0.0"), // dev
+        .package(url: "https://github.com/f-meloni/Rocket", from: "0.4.0"), // dev
+        .package(url: "https://github.com/jpsim/Yams.git", from: "1.0.0"), // dev
     ],
     targets: [
+        .target(name: "Danger-Swift", dependencies: ["Danger", "Yams"]), // dev
         .target(name: "Danger", dependencies: ["ShellOut", "OctoKit", "Logger"]),
         .target(name: "RunnerLib", dependencies: ["Logger", "ShellOut"]),
         .target(name: "Runner", dependencies: ["RunnerLib", "MarathonCore", "Logger"]),
@@ -49,20 +52,15 @@ let package = Package(
             ],
         ],
         "rocket": [
-            "steps": [
-                ["script": ["content": "make docs"]],
-                ["script": ["content": "Scripts/update_makefile.sh"]],
-                ["script": ["content": "Scripts/update_danger_version.sh"]],
-                ["script": ["content": "Scripts/update_changelog.sh"]],
-                "hide_dev_dependencies",
-                ["git_add": ["paths": ["Makefile", "CHANGELOG.md", "Sources/Runner/main.swift", "Documentation", "Package.swift"]]],
-                ["commit": ["no_verify": true]],
-                "tag",
-                "unhide_dev_dependencies",
-                ["git_add": ["paths": ["Package.swift"]]],
-                ["commit": ["message": "Unhide dev dependencies", "no_verify": true]],
+            "before": [
+                "make docs",
+                "Scripts/update_makefile.sh",
+                "Scripts/update_danger_version.sh",
+                "Scripts/update_changelog.sh",
+            ],
+            "after": [
                 "push",
-                ["script": ["content": "Scripts/create_homebrew_tap.sh"]],
+                "Scripts/create_homebrew_tap.sh",
             ],
         ],
     ])