+++ Sources/Runner/Commands/RunDangerJS.swift
@@ -3,7 +3,16 @@ import Logger
 import RunnerLib
 
 func runDangerJSCommandToRunDangerSwift(_ command: DangerCommand, logger: Logger) throws -> Int32 {
-    let dangerJS = try getDangerCommandPath(command, logger: logger)
+    let dangerJS = try getDangerCommandPath(logger: logger)
+    let dangerJSVersion = try DangerJSVersionFinder.findDangerJSVersion(dangerJSPath: dangerJS)
+
+    guard dangerJSVersion.compare(MinimumDangerJSVersion, options: .numeric) != .orderedAscending else {
+        logger.logError("The installed danger-js version is below the minimum supported version",
+                        "Current version = \(dangerJSVersion)",
+                        "Minimum supported version = \(MinimumDangerJSVersion)",
+                        separator: "\n")
+        exit(1)
+    }
 
     let proc = Process()
     proc.environment = ProcessInfo.processInfo.environment
@@ -20,7 +29,7 @@ func runDangerJSCommandToRunDangerSwift(_ command: DangerCommand, logger: Logger
         dangerSwiftCommand = ".build/debug/danger-swift"
     }
 
-    proc.arguments = ["--process", dangerSwiftCommand, "--passURLForDSL"] + unusedArgs
+    proc.arguments = [command.rawValue, "--process", dangerSwiftCommand, "--passURLForDSL"] + unusedArgs
 
     let standardOutput = FileHandle.standardOutput
     proc.standardOutput = standardOutput