+++ Sources/RunnerLib/CliArgs/CliArgsParser.swift
@@ -1,10 +1,3 @@
-//
-//  CliArgsParser.swift
-//  Danger
-//
-//  Created by Franco Meloni on 12/11/2018.
-//
-
 import Foundation
 
 public final class CliArgsParser {